backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-jose[cryptography]==3.3.0
google-cloud-firestore==2.19.0
webauthn==2.1.0
httpx==0.28.1
